Move header struct declaration to separate file
This allows building other (C-based) tools that use the same definitions
that Perl is using, in a consistent way.
